Make the callbacks more clear to understand by directly injectin them into their correct place without passing the dict to each function what would cause some confusion in some cases
